KeyOver1/requirements.txt
psycopg2-binary
pandas
numpy
scikit-learn
joblib
